fix: url of UploadLogs.gif and don't require space

+++ src/commands/LogsCommand.ts
@@ -10,7 +10,7 @@ class LogsCommand extends BaseCommand {
      * The pattern to match against. If the message matches this pattern then we will respond to it with the action
      * method.
      */
-    pattern = /^!logs\s/;
+    pattern = /^!logs/;
 
     /**
      * The function that should be called when the event is fired.
@@ -22,21 +22,23 @@ class LogsCommand extends BaseCommand {
 
         const sentMessage = await message.channel.send(
             `In order to help you${userPre}${user}, we need some logs. Please see ` +
-                'https://enderman.atlcdn.net/UploadLogs.gif on how to generate the link. Please make sure that you ' +
+                'https://cdn.atlcdn.net/UploadLogs.gif on how to generate the link. Please make sure that you ' +
                 'press the button after the error/issue occurs.  Once done please paste the link here. If the logs ' +
                 "don't upload or this is an issue with a server, please upload your logs to https://pastebin.com/ " +
                 'and give us the link.',
         );
 
-        message.delete();
+        if (message.channel.type !== 'dm') {
+            message.delete();
 
-        // delete message after 24 hours
-        sentMessage.delete({ timeout: 60 * 60 * 24 * 1000 });
+            // delete message after 24 hours
+            sentMessage.delete({ timeout: 60 * 60 * 24 * 1000 });
 
-        await sentMessage.react('🇱');
-        await sentMessage.react('🇴');
-        await sentMessage.react('🇬');
-        sentMessage.react('🇸');
+            await sentMessage.react('🇱');
+            await sentMessage.react('🇴');
+            await sentMessage.react('🇬');
+            await sentMessage.react('🇸');
+        }
     }
 }
